Experimental modification of forward analysis and rewrite
This implementation discards graph entries passed in by the user
and uses factbase passed in to analysis to get the list of known
entry nodes. Note: this forces user to actually pass in facts
about all entry points he wants analysed or else they will be
discarded in the rewritten graph. Previously bottom was assumed
implicitly.